requirements-cu111.txt
--find-links https://pytorch-geometric.com/whl/torch-1.8.0+cu111.html
torch_scatter==2.0.8
--find-links https://pytorch-geometric.com/whl/torch-1.8.0+cu111.html
torch-sparse==0.6.12
--find-links https://pytorch-geometric.com/whl/torch-1.8.0+cu111.html
torch-cluster==1.5.9
--find-links https://pytorch-geometric.com/whl/torch-1.8.0+cu111.html
torch-spline-conv==1.2.1
torch-geometric==2.0.1
ase==3.20.1
chardet==3.0.4
cycler==0.10.0
future==0.18.2
googledrivedownloader==0.4
h5py==2.10.0
idna==2.10
isodate==0.6.0
jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
llvmlite==0.34.0
markupsafe==1.1.1
matplotlib==3.3.1
networkx==2.5
numba==0.51.0
pandas==1.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
rdflib==5.0.0
requests==2.24.0
scikit-learn==0.23.2
threadpoolctl==2.1.0
tqdm==4.48.2
urllib3==1.25.10
opencv-python
opencv-contrib-python
tensorboardX
loguru
tensorboard
scikit-image
py-lz4framed
setproctitle